// src/models/pagination.model.ts
export class PaginationModel {
    page: number;
    role: string;
    limit: number;
    status: number;
    tag: string;
    company: string;
    driver: string;
    assignedStatus: number;
    shipment: string;
    keyword: string;
    startDate: string;
    endDate: string;
    user: string;
    sortBy: string;
    sortValue: string;

}